Introduce the AccountService and dependent models
To facilite Account Switching an abstraction layer needs to be placed on top of StorageService, UserService, and any other singletons that hold user data.
This abstraction, the AccountService, must maintain a context for any amount of logged in Bitwarden accounts, and allow for manipulating data stored in memory and on disk for the currently active account.
* Create an abstraction and implementation of the AccountService that exposes methods for adding switching, and removing accounts, as well as saving, fetching, and removing data stored in memory or on disk for the active user.
* Create an AccountStorageKey enum to centralize various storage keys used across services. In addition to being a cleaner solution this allows for reusing the same keys between the in memory and on disk storage mechanisms.
* Create an Account object to represent a signed in user account.